notifications/Notification-requestPermission-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Notifications Test: Notification - requestPermission (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<link rel="help" title="4.3 Permission" href="http://www.w3.org/TR/notifications/#permission">
<meta name="flags" content="">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <div id="log"></div>
  <script>
    test(function() {
      var permission_granted;
      Notification.requestPermission(function(permission) {
        permission_granted = permission;
      })
      assert_in_array(permission_granted, ["default", "denied", "granted"], "notification permission");
    }, "Test checks that notification the value of permission is granted");
  </script>
</body>
